Removed Makefile.mmpbsa since it is unnecessary after using automake
Started the integration of Sander into MMPBSA. Right now it forks and
controls sander as its own process. BOINC functions remain in MMPBSA,
which will handle reporting sander's progress, as well as controlling
sander based on the client's wishes. This commit marks the prototype
of this integration. Testing and refinement are needed.